+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Compatibility:
 
 * Fix `mkmf.rb` to use host's `RbConfig` for `$ruby` when `--target-rbconfig` is used (#4097, @panozzaj).
 * Implement `Fiber.[]`, `Fiber.[]=`, and `Fiber#{storage,storage=}` (#4007, @nirvdrum).
+* Fix `Symbol#inspect` to handle incompatible encodings (#4106, @eregon).
 
 Performance: